style: :construction: Modal window ready with its respective information, missing action to send to POST
We are working on opening a modal screen for each scholarship that exists, in addition to having specific information about these. Finally, we would need to find a way to make the payment when the modal window opens, we are trying to use a form or create a  POST request.